+++ Game/src/entitys/entity.java
@@ -7,11 +7,11 @@ public abstract class entity {
 	public float x;
 	public float y;
 	public float health;
-	public float damage;
+	public double damage;
 	protected String type;
 	protected float w = 32;
 	protected float h = 32;
-	public entity(float x, float y, String type, float health, float damage)
+	public entity(float x, float y, String type, float health, double damage)
 	{
 		this.x = x;
 		this.y = y;
